Moved removeFakeNodes to the html-parser so nobody gets fake nodes in their resolved document ASTs.
